수정
수정

+++ Assets/02_Script/Jang/WindowManamer/Alcohol.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class Alcohol : ScriptableObject
 {
 
-    public float sell;
+    public int sell;
     public float buff_Speed;
     public float buff_Defence;
